refactor: extract Plot2D and PlotMesh into plot2d/ subpackage
