+++ anyplotlib/plot2d/__init__.py
@@ -1,0 +1,4 @@
+from anyplotlib.plot2d._plot2d import Plot2D
+from anyplotlib.plot2d._plotmesh import PlotMesh
+
+__all__ = ["Plot2D", "PlotMesh"]
